clear endpoint FIXED: 10 tests pass

+++ server/src/main/java/service/LogoutService.java
@@ -3,7 +3,6 @@ package service;
 import dataaccess.AuthDAO;
 import dataaccess.DataAccessException;
 import dataaccess.ResponseException;
-import dataaccess.UserDAO;
 import model.AuthData;
 
 public class LogoutService {
@@ -16,6 +15,13 @@ public class LogoutService {
 
     public LogoutResult logout(LogoutRequest request) throws ResponseException {
         try {
+
+            var authToken = authDAO.getAuth(request.authToken());
+
+            if (authToken == null) {
+                throw new ResponseException(401, "Error: unauthorized");
+            }
+
             authDAO.deleteAuth(request.authToken());
 
             return new LogoutResult("logged out");
